requirements.txt
requests==2.31.0
python-dotenv==1.0.0
chromadb==0.4.0
neo4j==5.0.0
stripe==13.0.0
docker==6.1.3
kubernetes==27.2.0
pytest==7.4.0
pydantic==2.0.0
fastapi==0.100.0
uvicorn==0.23.0
